refactor: optimize folder structure
