+++ lib/api/DeviceAuthInfo.dart
