+++ lib/api/routes/DevicesRoute.dart
