+++ lib/api/routes/StocksRoute.dart
